api/requirements.txt
aiohttp
aioredis
asyncio
black
Celery
celery[redis]
celery[sqlalchemy]
coloredlogs
crewai
crewai-tools
crewai[tools]
databases[postgresql]
fastapi[standard]
jira
langchain
langchain-experimental
langchain-openai
langchain_community
langchain_core
langgraph
langserve[all]
logfire
openai
pandas
prisma
psycopg2
pydantic
pytest
pytest-asyncio
python-dateutil
python-dotenv
pyyaml
redis
requests
rq
starlette
sqlalchemy
sse-starlette
tqdm
uvicorn